[IA64] fix make world
When the arch/ia64/Makefile was cloned from the Linux tree, the install
target was changed to remove a dependency on vmlinux.gz.  That makes
sense, but the invocation of install.sh in that target still used $<
to refer to the dependency that had been removed.  Hence, install.sh
would fail, and make world would stop prematurely.

Signed-off-by: Al Stone <ahs3@fc.hp.com>